requirement.txt
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
cryptography==2.9
idna==2.9
numpy==1.18.2
Pillow==7.1.1
pycparser==2.20
pynput==1.6.8
pyobjc-core==6.2
pyobjc-framework-Cocoa==6.2
pyobjc-framework-Quartz==6.2
requests==2.23.0
scipy==1.4.1
six==1.14.0
sounddevice==0.3.15
urllib3==1.25.8